requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
Deprecated==1.2.13
Django==4.0.2
django-cors-headers==3.11.0
django-filter==21.1
django-oauth-toolkit==1.7.0
djangorestframework==3.13.1
idna==3.3
jwcrypto==1.0
oauthlib==3.2.0
psycopg2-binary==2.9.3
pycparser==2.21
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.8
wrapt==1.13.3
